+++ CHANGELOG.md
@@ -5,6 +5,7 @@ CHANGE LOG
 ## V2.0 Beta 1 (Upcoming - July 2014)
 
 * Upgrade to Laravel 4.2
+* Added minification filter
 
 
 ## V1.1.1 (30/04/2014)